third_party/blink/web_tests/fast/css/zoom-media-queries-resolution-expected.txt

PASS matchMedia("(resolution: " + initialResolution + "dppx)").matches is true
PASS matchMedia("(-webkit-device-pixel-ratio: " + initialResolution + ")").matches is true
PASS matchMedia("(resolution: " + initialResolution * 1.2 + "dppx)").matches is true
PASS matchMedia("(-webkit-device-pixel-ratio: " + initialResolution * 1.2 + ")").matches is true
PASS matchMedia("(resolution: " + initialResolution / 1.2 + "dppx)").matches is true
PASS matchMedia("(-webkit-device-pixel-ratio: " + initialResolution / 1.2 + ")").matches is true
PASS successfullyParsed is true

TEST COMPLETE
